requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiogram3-calendar==0.1.2
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
h3==3.7.6
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
numpy==1.26.3
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyswisseph==2.10.3.2
pytz==2023.3.post1
timezonefinder==6.2.0
typing_extensions==4.9.0
yarl==1.9.4